% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotdataplane_operations.R
\name{delete_thing_shadow}
\alias{delete_thing_shadow}
\title{Deletes the thing shadow for the specified thing}
\usage{
delete_thing_shadow(
  thingName = "string"
)
}
\arguments{
\item{thingName}{[required] The name of the thing.}
}
\description{
Deletes the thing shadow for the specified thing.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/iot/latest/developerguide/API_DeleteThingShadow.html}{DeleteThingShadow} in the \emph{AWS IoT Developer Guide}.
}
\examples{

}
