% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotdataplane_package.R
\docType{package}
\name{paws.iotdataplane-package}
\alias{paws.iotdataplane}
\alias{paws.iotdataplane-package}
\title{paws.iotdataplane: AWS IoT Data Plane}
\description{
AWS IoT-Data enables secure, bi-directional
communication between Internet-connected things (such as sensors,
actuators, embedded devices, or smart appliances) and the AWS cloud.
It implements a broker for applications and things to publish messages
over HTTP (Publish) and retrieve, update, and delete thing shadows. A
thing shadow is a persistent representation of your things and their
state in the AWS cloud.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
