% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotdataplane_operations.R
\name{publish}
\alias{publish}
\title{Publishes state information}
\usage{
publish(
  topic = "string",
  qos = 123,
  payload = raw
)
}
\arguments{
\item{topic}{[required] The name of the MQTT topic.}

\item{qos}{The Quality of Service (QoS) level.}

\item{payload}{The state information, in JSON format.}
}
\description{
Publishes state information.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/iot/latest/developerguide/protocols.html#http}{HTTP Protocol} in the \emph{AWS IoT Developer Guide}.
}
\examples{

}
