% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotdataplane_operations.R
\name{update_thing_shadow}
\alias{update_thing_shadow}
\title{Updates the thing shadow for the specified thing}
\usage{
update_thing_shadow(
  thingName = "string",
  payload = raw
)
}
\arguments{
\item{thingName}{[required] The name of the thing.}

\item{payload}{[required] The state information, in JSON format.}
}
\description{
Updates the thing shadow for the specified thing.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/iot/latest/developerguide/API_UpdateThingShadow.html}{UpdateThingShadow} in the \emph{AWS IoT Developer Guide}.
}
\examples{

}
